update extension terms and guidelines page
